penyempurnaan hak akses siswa

+++ app/Http/Controllers/RuanganController.php
@@ -14,8 +14,12 @@ class RuanganController extends Controller
 {
     public function index($lantai = 'Lantai 1')
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha') {
-            return redirect()->route('showLantai', ['lantai' => 'Lantai 1']);
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha') {
+                return redirect()->route('showLantai', ['lantai' => 'Lantai 1']);
+            } else {
+                return back();
+            }
         } else {
             return back();
         }
@@ -26,14 +30,18 @@ class RuanganController extends Controller
      */
     public function showLantai($lantai)
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha') {
-            $title = "Ruangan";
-            $dataRuangan = Ruangan::where('lantai', $lantai)->get();
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha') {
+                $title = "Ruangan";
+                $dataRuangan = Ruangan::where('lantai', $lantai)->get();
 
-            confirmDelete();
+                confirmDelete();
 
-            // Redirect
-            return view('dashboard.Sarpra.DataRuangan', compact('dataRuangan', 'lantai', 'title'));
+                // Redirect
+                return view('dashboard.Sarpra.DataRuangan', compact('dataRuangan', 'lantai', 'title'));
+            } else {
+                return back();
+            }
         } else {
             return back();
         }
@@ -44,12 +52,16 @@ class RuanganController extends Controller
      */
     public function create(Request $request)
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha') {
-            $title = "Tambah Ruangan";
-            $lantai = $request->query('lantai', 'Lantai 1');
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha') {
+                $title = "Tambah Ruangan";
+                $lantai = $request->query('lantai', 'Lantai 1');
 
-            // Redirect
-            return view('dashboard.Sarpra.createRuangan', compact('title', 'lantai'));
+                // Redirect
+                return view('dashboard.Sarpra.createRuangan', compact('title', 'lantai'));
+            } else {
+                return back();
+            }
         } else {
             return back();
         }
@@ -61,39 +73,43 @@ class RuanganController extends Controller
      */
     public function store(Request $request)
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha') {
-            $messages = [
-                'nama.required' => 'Nama harus diisi.',
-                'nama.string' => 'Nama harus berupa teks.',
-                'nama.max' => 'Nama tidak boleh lebih dari 100 karakter.',
-                'deskripsi.string' => 'Deskripsi harus berupa teks.',
-                'lantai.required' => 'Lantai harus dipilih.',
-                'lantai.in' => 'Pilihan lantai tidak valid. Pilih antara: Lantai 1, Lantai 2, atau Lantai 3.',
-            ];
-            // Validasi input
-            $validator = Validator::make($request->all(), [
-                'nama' => 'required|string|max:100',
-                'deskripsi' => 'nullable|string',
-                'lantai' => 'required|in:Lantai 1,Lantai 2,Lantai 3',
-            ], $messages);
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha') {
+                $messages = [
+                    'nama.required' => 'Nama harus diisi.',
+                    'nama.string' => 'Nama harus berupa teks.',
+                    'nama.max' => 'Nama tidak boleh lebih dari 100 karakter.',
+                    'deskripsi.string' => 'Deskripsi harus berupa teks.',
+                    'lantai.required' => 'Lantai harus dipilih.',
+                    'lantai.in' => 'Pilihan lantai tidak valid. Pilih antara: Lantai 1, Lantai 2, atau Lantai 3.',
+                ];
+                // Validasi input
+                $validator = Validator::make($request->all(), [
+                    'nama' => 'required|string|max:100',
+                    'deskripsi' => 'nullable|string',
+                    'lantai' => 'required|in:Lantai 1,Lantai 2,Lantai 3',
+                ], $messages);
 
-            // Jika validasi gagal, kembalikan pesan error
-            if ($validator->fails()) {
-                return redirect()->back()->withErrors($validator)->withInput();
+                // Jika validasi gagal, kembalikan pesan error
+                if ($validator->fails()) {
+                    return redirect()->back()->withErrors($validator)->withInput();
+                }
+
+                // Simpan data ke database
+                $ruangan = new Ruangan();
+                $ruangan->nama = $request->input('nama');
+                $ruangan->deskripsi = $request->input('deskripsi');
+                $ruangan->lantai = $request->input('lantai');
+                $ruangan->save();
+
+                // Sweet alert
+                Alert::success('Berhasil Ditambahkan', 'Ruangan berhasil ditambahkan.');
+
+                // Redirect
+                return redirect()->route('showLantai', $ruangan->lantai);
+            } else {
+                return back();
             }
-
-            // Simpan data ke database
-            $ruangan = new Ruangan();
-            $ruangan->nama = $request->input('nama');
-            $ruangan->deskripsi = $request->input('deskripsi');
-            $ruangan->lantai = $request->input('lantai');
-            $ruangan->save();
-
-            // Sweet alert
-            Alert::success('Berhasil Ditambahkan', 'Ruangan berhasil ditambahkan.');
-
-            // Redirect
-            return redirect()->route('showLantai', $ruangan->lantai);
         } else {
             return back();
         }
@@ -111,14 +127,18 @@ class RuanganController extends Controller
      */
     public function edit(string $id)
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha') {
-            // Find ruangan by id
-            $ruangan = Ruangan::query()->find((integer) $id);
-            $lantai = $ruangan->lantai;
-            $title = "Edit Ruangan";
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha') {
+                // Find ruangan by id
+                $ruangan = Ruangan::query()->find((integer) $id);
+                $lantai = $ruangan->lantai;
+                $title = "Edit Ruangan";
 
-            // Redirect
-            return view('dashboard.Sarpra.EditRuangan', data: compact(['ruangan', 'title', 'lantai']));
+                // Redirect
+                return view('dashboard.Sarpra.EditRuangan', data: compact(['ruangan', 'title', 'lantai']));
+            } else {
+                return back();
+            }
         } else {
             return back();
         }
@@ -129,39 +149,43 @@ class RuanganController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha') {
-            // Validasi input
-            $messages = [
-                'nama.required' => 'Nama harus diisi.',
-                'nama.string' => 'Nama harus berupa teks.',
-                'nama.max' => 'Nama tidak boleh lebih dari 100 karakter.',
-                'deskripsi.string' => 'Deskripsi harus berupa teks.',
-                'lantai.required' => 'Lantai harus dipilih.',
-                'lantai.in' => 'Pilihan lantai tidak valid. Pilih antara: Lantai 1, Lantai 2, atau Lantai 3.',
-            ];
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha') {
+                // Validasi input
+                $messages = [
+                    'nama.required' => 'Nama harus diisi.',
+                    'nama.string' => 'Nama harus berupa teks.',
+                    'nama.max' => 'Nama tidak boleh lebih dari 100 karakter.',
+                    'deskripsi.string' => 'Deskripsi harus berupa teks.',
+                    'lantai.required' => 'Lantai harus dipilih.',
+                    'lantai.in' => 'Pilihan lantai tidak valid. Pilih antara: Lantai 1, Lantai 2, atau Lantai 3.',
+                ];
 
-            $validator = Validator::make($request->all(), [
-                'nama' => 'required|string|max:100',
-                'deskripsi' => 'nullable|string',
-                'lantai' => 'required|in:Lantai 1,Lantai 2,Lantai 3',
-            ],$messages);
+                $validator = Validator::make($request->all(), [
+                    'nama' => 'required|string|max:100',
+                    'deskripsi' => 'nullable|string',
+                    'lantai' => 'required|in:Lantai 1,Lantai 2,Lantai 3',
+                ], $messages);
 
-            if ($validator->fails()) {
-                return redirect()->back()->withErrors($validator)->withInput();
+                if ($validator->fails()) {
+                    return redirect()->back()->withErrors($validator)->withInput();
+                }
+
+                // First find by id
+                $ruangan = Ruangan::findOrFail($id);
+
+                // Update data use fill
+                $ruangan->fill($request->only(['nama', 'deskripsi', 'lantai']));
+                $ruangan->save();
+
+                // Sweet alert
+                Alert::success('Perubahan Berhasil', 'Ruangan berhasil diubah.');
+
+                // Redirect
+                return redirect()->route('showLantai', $ruangan->lantai);
+            } else {
+                return back();
             }
-
-            // First find by id
-            $ruangan = Ruangan::findOrFail($id);
-
-            // Update data use fill
-            $ruangan->fill($request->only(['nama', 'deskripsi', 'lantai']));
-            $ruangan->save();
-
-            // Sweet alert
-            Alert::success('Perubahan Berhasil', 'Ruangan berhasil diubah.');
-
-            // Redirect
-            return redirect()->route('showLantai', $ruangan->lantai);
         } else {
             return back();
         }
@@ -173,19 +197,23 @@ class RuanganController extends Controller
      */
     public function destroy(string $id)
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha') {
-            // Find by id
-            $findRuangan = Ruangan::findOrFail((integer) $id);
-            $lantai = $findRuangan->lantai;
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha') {
+                // Find by id
+                $findRuangan = Ruangan::findOrFail((integer) $id);
+                $lantai = $findRuangan->lantai;
 
-            // Delete ruangan
-            $findRuangan->delete();
+                // Delete ruangan
+                $findRuangan->delete();
 
-            // Sweet alert
-            Alert::success('Berhasil Dihapus', 'Ruangan berhasil dihapus.');
+                // Sweet alert
+                Alert::success('Berhasil Dihapus', 'Ruangan berhasil dihapus.');
 
-            // Redirect
-            return redirect()->route('showLantai', $lantai);
+                // Redirect
+                return redirect()->route('showLantai', $lantai);
+            } else {
+                return back();
+            }
         } else {
             return back();
         }
